+++ app/Http/Controllers/EventDetail.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Http\Controllers;
+
+
+use App\Models\Event;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+
+class EventDetail extends Controller
+{
+
+    public function index(Event $event)
+    {
+        if (auth()->guest() &&
+            (!$event->enabled || $event->logged_in_only)) {
+            throw new ModelNotFoundException();
+        }
+        return view("events.event-detail", ["event" => $event]);
+    }
+}